Update to team folder and change export default in indexJS file under submission
