+++ src/Page/Submission/Team1/Team1.js
@@ -1,19 +1,19 @@
-// Annabel Qi
-// Ordencia Wu
-// Esther Jin
-// Mandy Cao
+// Sherman Grewal
+// Oliver Lingertat
+// Luncy Cheng
+// Navya Mehta
 
 import React, { Component } from 'react';
 // import Highcharts from 'highcharts';
 // import HighchartsReact from 'highcharts-react-official';
 // import { Jumbotron } from 'reactstrap';
 
-class Team5 extends Component { 
+class Team1 extends Component { 
   render() {
     return (
-      <div className='container'>Team 
-       </div>
+      <div className='container'>Team 1</div>
     )
   }
 }
-export default Team5;
+
+export default Team1;
