+++ src/Page/Submission/Team10/Team10.js
@@ -1,17 +1,17 @@
-// Jason Huang
-// Proannoy Lal
+// Yixin Li
+// Lexi Gao
 
 import React, { Component } from 'react';
 // import Highcharts from 'highcharts';
 // import HighchartsReact from 'highcharts-react-official';
 // import { Jumbotron } from 'reactstrap';
 
-class Team9 extends Component { 
+class Team10 extends Component { 
   render() {
     return (
-      <div className='container'>Team 
-       </div>
+      <div className='container'>Team 9</div>
     )
   }
 }
-export default Team9;
+
+export default Team10;
