+++ src/Page/Submission/Team3/Team3.js
@@ -1,17 +1,19 @@
-// Anthony Ilersich
-// Evelyn Law
-// Lilian Chen
-// Aaron Kong
+// Yingying Huang
+// Even Yu
+// Audrey Sin Fai Lam
+// Vaanie Kathirkamar
+
 import React, { Component } from 'react';
 // import Highcharts from 'highcharts';
 // import HighchartsReact from 'highcharts-react-official';
 // import { Jumbotron } from 'reactstrap';
 
-class Team2 extends Component { 
+class Team3 extends Component { 
   render() {
     return (
-      <div className='container'>Team 2</div>
+      <div className='container'>Team 3</div>
     )
   }
 }
-export default Team2;
+ 
+export default Team3;
