+++ src/Page/Submission/Team4/Team4.js
@@ -1,19 +1,19 @@
-// Estelle Wang
-// Lia Yi
-// Phoebe Shi
-// David Chung
+// Samuel Chen
+// Steve Sam-Soon
+// Dhayria Rao
+// Fox Mutual
 
 import React, { Component } from 'react';
 // import Highcharts from 'highcharts';
 // import HighchartsReact from 'highcharts-react-official';
 // import { Jumbotron } from 'reactstrap';
 
-class Team7 extends Component { 
+class Team4 extends Component { 
   render() {
     return (
-      <div className='container'>Team 
-       </div>
+      <div className='container'>Team 4</div>
     )
   }
 }
-export default Team7;
+
+export default Team4;
