+++ src/Page/Submission/Team6/Team6.js
@@ -1,19 +1,19 @@
-// Yingying Huang
-// Even Yu
-// Audrey Sin Fai Lam
-// Vaanie Kathirkamar
+// Yiqing Xia
+// Xinyue Zhao
+// Amy Li
+// Felicia Shang
 
 import React, { Component } from 'react';
 // import Highcharts from 'highcharts';
 // import HighchartsReact from 'highcharts-react-official';
 // import { Jumbotron } from 'reactstrap';
 
-class Team3 extends Component { 
+class Team6 extends Component { 
   render() {
     return (
-      <div className='container'>Team 3:   
-       </div>
+      <div className='container'>Team 6</div>
     )
   }
 }
-export default Team3;
+
+export default Team6;
