+++ src/Page/Submission/Team7/Team7.js
@@ -1,19 +1,19 @@
-// Yiqing Xia
-// Xinyue Zhao
-// Amy Li
-// Felicia Shang
+// Estelle Wang
+// Lia Yi
+// Phoebe Shi
+// David Chung
 
 import React, { Component } from 'react';
 // import Highcharts from 'highcharts';
 // import HighchartsReact from 'highcharts-react-official';
 // import { Jumbotron } from 'reactstrap';
 
-class Team6 extends Component { 
+class Team7 extends Component { 
   render() {
     return (
-      <div className='container'>Team 
-       </div>
+      <div className='container'>Team 7</div>
     )
   }
 }
-export default Team6;
+
+export default Team7;
