+++ src/Page/Submission/Team8/Team8.js
@@ -1,18 +1,19 @@
-// Yixin Li
-// Lexi Gao
-
+// Yiming Shi
+// Allen Wu
+// Chenkai Zhu
+// Ayman Kamran
 
 import React, { Component } from 'react';
 // import Highcharts from 'highcharts';
 // import HighchartsReact from 'highcharts-react-official';
 // import { Jumbotron } from 'reactstrap';
 
-class Team10 extends Component { 
+class Team8 extends Component { 
   render() {
     return (
-      <div className='container'>Team 
-       </div>
+      <div className='container'>Team 8</div>
     )
   }
 }
-export default Team10;
+
+export default Team8;
