+++ src/Page/Submission/Team9/Team9.js
@@ -1,19 +1,17 @@
-// Yiming Shi
-// Allen Wu
-// Chenkai Zhu
-// Ayman Kamran
+// Jason Huang
+// Proannoy Lal
 
 import React, { Component } from 'react';
 // import Highcharts from 'highcharts';
 // import HighchartsReact from 'highcharts-react-official';
 // import { Jumbotron } from 'reactstrap';
 
-class Team8 extends Component { 
+class Team9 extends Component { 
   render() {
     return (
-      <div className='container'>Team 
-       </div>
+      <div className='container'>Team 9</div>
     )
   }
 }
-export default Team8;
+
+export default Team9;
